Changed the structure of string struct. 	The struct now contains 3 values 1) lenght of string 2) pointer to the "end" node 3) pointer to the "start" node. 	This allowed us to remove the anomallies relatedd to the special case of i being 0 and treat all the characters equally. 	The "start" node has next set to end during initialising the string and similarly the "end" node has prev set to start. This changes as you add more characters. 	The process of adding a character changed and the special case off adding characters in between no-longer needs a special attention adding a character anywhere always follow the same algorithm 	The value field of start and end nodes can easily be changed by the programmers using "str.end/start->value =" this supports customisation during strprint()
	Added a new feature of detecting left and right arrows and supports insertion both visually and in backgroud.
	The string now ends reading as soon as enter key is pressed.

The writer can now take a new approach on storing info by storing lines. This will help in managing the data time efficiently as to support insertion the whole text would have to be
printed again and again. So instead a new struct line is introduced that is like a node but for strings instead of characters. Just like string is represented as collection of nodes the
test data type is a collection of lines.

TODO:
1)	More functions related to lines for things like supporting vertical motion in the text. gettxt(), printtxt(), freetxt() ....